timescale-fastapi/requirements.txt
python-dotenv
asyncpg
loguru
fastapi
uvicorn
requests
redis.asyncio
redis-timeseries
aioredis
